wip: useToken hook
