+++ src/components/client-root.tsx
@@ -15,7 +15,7 @@ export default function ClientRoot({ children }: { children: ReactNode }) {
 
   const {
     accessToken,
-    clearAccessToken,
+    setAccessToken,
     user,
     watching,
     addWatching,
@@ -38,7 +38,7 @@ export default function ClientRoot({ children }: { children: ReactNode }) {
     <div id="root" className="flex flex-col h-screen">
       <Header
         accessToken={accessToken}
-        clearAccessToken={clearAccessToken}
+        setAccessToken={setAccessToken}
         user={user}
         addWatching={addWatching}
         playerView={playerView}
