+++ src/hooks/path.tsx
@@ -14,6 +14,7 @@ export function usePathSync(
   playerView: PlayerView,
   setPlayerView: (view: PlayerView) => void
 ) {
+  console.log('test usePathSync');
   const pathname = usePathname();
   const searchParams = useSearchParams();
   const isSyncing = useRef(false);
@@ -21,6 +22,7 @@ export function usePathSync(
 
   // path to state
   useEffect(() => {
+    console.log('test2');
     // safegaurd
     if (isSyncing.current) return;
 
@@ -52,6 +54,7 @@ export function usePathSync(
 
   // state to path
   useEffect(() => {
+    console.log('test3');
     if (isSyncing.current) return;
 
     const newPath = `/${order.join('/')}`;
